logout using post only

+++ handlers/logout.go
@@ -10,6 +10,11 @@ import (
 
 func CreateLogoutHandler(db *pg.DB) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
+		if req.Method != "POST" {
+			jsonhttp.ErrorResponse(w, "use POST", nil, 400)
+			return
+		}
+
 		claims := req.Context().Value("claims").(*jwtutils.Claims)
 
 		err := dao.DeleteJwtToken(db, claims.TokenId)
